chatgpt button added to product form -start working on functionality
