+++ admin/src/components/index.js
@@ -56,7 +56,6 @@ export { default as ToggleConfigButton } from './ToggleConfigButton';
 export { default as ColorPicker } from './ColorPicker';
 export { default as CustomModal } from './CustomModal';
 export { default as LimitCharacter } from './LimitCharacter';
-export { default as ChatGPTButton } from './ChatGPTButton';
 // export { default as DeleteElement } from './DeleteElement';
 // export { default as PostFilterButton } from './PostFilterButton';
 // export { default as PostFilterForm } from './PostFilterForm';
